--- tocount/linear/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Linear token estimator."""
-
-from .estimator import linear_tokens_estimator
