+++ tocount/tiktoken_r50/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""TikToken R50K token estimator."""
+
+from .functions import linear_tokens_estimator
